% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_interactions.R
\name{Assign_Celltypes}
\alias{Assign_Celltypes}
\title{Cell type Assignment (Doublet deconvolution)}
\usage{
Assign_Celltypes(
  refObj,
  rawObj,
  NBres,
  col.label,
  nPC = 20,
  thres = 0.98,
  OutPath
)
}
\arguments{
\item{refObj}{Reference set; A fully processed (from QC to annotation) Seurat object}

\item{rawObj}{A Seurat object of raw input expression matrix}

\item{NBres}{Returned object from function 'DefineDoublets', where putative doublets called from doublet meta-caller and
their information are included}

\item{col.label}{Name of the column containing cell type label in the reference set}

\item{nPC}{The number of PCs to use for label transferring. Default to 20}

\item{thres}{Threshold value in range (0, 1), which is used to filter maximal cell type prediction score. Default to 0.98}

\item{OutPath}{Path to the output directory under which log files of defining doublet set are saved.
(Path used as the parameter PATH_out in previous step of Define_doublets().)}
}
\description{
Transferring cell types from reference set to doublet set using functions 'FindTransferAnchors' and 'TransferData' in Seurat pipeline.
}
