% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define_Doublets.R
\name{Bind_Doublet_Scores}
\alias{Bind_Doublet_Scores}
\title{Binding the doublet detection results}
\usage{
Bind_Doublet_Scores(type, nRun = 10, nSet, InPath)
}
\arguments{
\item{type}{Type of the input dataset, either 'raw' or 'simul'.}

\item{nRun}{An integer indicating the number of running iteration with different random seeds for each dataset. Default to 10}

\item{nSet}{The number of the dataset to run. Default to 1 for type 'raw' and 100 for type 'simul'.}

\item{InPath}{Path to the input directory in which input count matrix (for type='raw') or folders named 'simul_#' which contain
count matrices of doublet-simulated datasets (for type='simul') exist.}
}
\value{
A list of doublet scores calculated from three methods for each cell in the data.
}
\description{
Doublet scores calculated from three tools (Hybrid, scDblFinder, Scrublet ) are loaded.
For each tool and each dataset (either raw or simulated), detection results (.txt) are read and bound,
resulting in a list of data frame containing three doublet scores of each cell.
}
