% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define_Doublets.R
\name{DefineDoublets}
\alias{DefineDoublets}
\title{Executing all the functions for defining putative doublets in the data}
\usage{
DefineDoublets(nRun, nSet_s, nSet_a, PATH, PATH_out)
}
\arguments{
\item{nRun}{An integer indicating the number of running iteration with different random seeds for each dataset. Default to 10}

\item{nSet_s}{The number of simulated datasets. Default to 100.}

\item{nSet_a}{The number of actual input (raw) dataset. Default to 1}

\item{PATH}{Path to the input directory in which input count matrix (for type='raw') or folders named 'simul_#' which contain
count matrices of doublet-simulated datasets (for type='simul') exist.}

\item{PATH_out}{Path to the directory to save result of each step}
}
\description{
In this function, putative doublets exist in the data are defined.
}
