% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define_Doublets.R
\name{Doublet_calling}
\alias{Doublet_calling}
\title{Defining doublets using normalized doublet scores}
\usage{
Doublet_calling(SimulSet, RawSet)
}
\arguments{
\item{SimulSet}{A list containing normalized doublet scores of simulated datasets, resulted from function 'Scale_AvgScores'.}

\item{RawSet}{A data-frame containing normalized doublet scores of input (raw) dataset, resulted from function 'Scale_AvgScores'.}
}
\description{
For each doublet detection method, probability density functions (PDFs) of singlets and doublets in the simulated data
are calculated and then putative doublets in the input data are defined using Gaussian naive bayes classification, based on those PDFs.
}
