% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect_Doublets.R
\name{DoubletScoring_Hybrid}
\alias{DoubletScoring_Hybrid}
\title{Doublet scoring through Hybrid of scds, a R-based doublet detection program (Bais and Kostka, 2019)}
\usage{
DoubletScoring_Hybrid(InPath, OutPath, type, nSet, nRun = 10)
}
\arguments{
\item{InPath}{Path to the input directory in which input count matrix (for type='raw') or folders named 'simul_#' which contain
count matrices of doublet-simulated datasets (for type='simul') exist.}

\item{OutPath}{Path to the output directory in which the results files are saved.}

\item{type}{Type of the input dataset, either 'raw' or 'simul'.}

\item{nSet}{The number of dataset to run.}

\item{nRun}{An integer indicating the number of running iteration with different random seeds for each dataset. Default to 10}
}
\description{
Doublet score (probability) of each cell in the input data or simulated data is yielded by Hybrid of Scds.
}
