% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect_Doublets.R
\name{DoubletScoring_scDblFinder}
\alias{DoubletScoring_scDblFinder}
\title{Running scDblFinder (R-based doublet detection program)}
\usage{
DoubletScoring_scDblFinder(
  InPath,
  OutPath,
  type,
  nSet,
  nRun = 10,
  est.rate = NULL
)
}
\arguments{
\item{InPath}{Path to the input directory in which input count matrix (for type='raw') or folders named 'simul_#' which contain
count matrices of doublet-simulated datasets (for type='simul') exist.}

\item{OutPath}{Path to the output directory in which the results files are saved.}

\item{type}{Type of the input dataset, either 'raw' or 'simul'.}

\item{nSet}{The number of dataset to run.}

\item{nRun}{An integer indicating the number of running iteration with different random seeds for each dataset. Default to 10}

\item{est.rate}{Estimated doublet rates in your dataset.}
}
\description{
Doublet score (probability) of each cell in the input data or simulated data is yielded by scDblFinder.
}
