% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_interactions.R
\name{FindInteraction}
\alias{FindInteraction}
\title{Finding significant interactions}
\usage{
FindInteraction(DeconvObj, refObj, col.label, lower.frac = 0.1, OutPath)
}
\arguments{
\item{DeconvObj}{Result object from function 'Assign_Celltypes()'; a list containing doublet deconvolution result}

\item{refObj}{Reference set; A fully processed (from QC to annotation) Seurat object}

\item{col.label}{Name of the column containing cell type label in the reference set}

\item{lower.frac}{The minimal fraction (prediction score calculated from Cell type Assignment) of a cell-type
to be considered as a constituent cell of a droplet (multiplet).}

\item{OutPath}{Path to the output directory under which log files of defining doublet set are saved.
(Path used as the parameter PATH_out in previous step of Define_doublets().)}
}
\value{
A list containing cell type fraction of the data and information of all detected doublets including potential interactions
}
\description{
Fraction of each cell type in the data are calculated first to estimate the expected number of doublets.
Then, observed number of doublets are compared to the expected number for each cell type combination (interaction).
P-values are calculated by hypergeometric test (equivalent to one-sided Fisher's exact test) and adjusted through Bonferroni correction.
}
