% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_interactions.R
\name{FindTwoCelltypes}
\alias{FindTwoCelltypes}
\title{Assigning two representative cell types to each doublet}
\usage{
FindTwoCelltypes(predicted.df)
}
\arguments{
\item{predicted.df}{A data frame of cell type prediction scores for each putative doublet (cell barcode X cell types),
which is resulted from functions 'FindTransferAnchors' and 'TransferData'.}
}
\description{
For each doublet, two cell types of top 2 high prediction scores are assigned.
}
