% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infer_interactions.R
\name{Hypertest}
\alias{Hypertest}
\title{Hypergeometric test}
\usage{
Hypertest(data)
}
\arguments{
\item{data}{A data frame containing information of doublets including potential interactions}
}
\description{
Hypergeometric test
}
