% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{Lower_plot_Preprocess}
\alias{Lower_plot_Preprocess}
\title{Plotting gene/UMI counts for quality control}
\usage{
Lower_plot_Preprocess(rawObj, expMtx, OutPath, project.id = "Project")
}
\arguments{
\item{rawObj}{A Seurat object of input gene expression matrix. This is not required if 'expMtx' is supplied.}

\item{expMtx}{Input gene expression count matrix. This is not required if 'rawObj' is supplied.}

\item{OutPath}{Path to the output directory under which subdirectories for saving the log files are created.}

\item{project.id}{A string which is used for 'orig.ident' in Seurat object and library name. Default to 'Project'.}
}
\description{
Jitter plots and histograms of gene and UMI counts are created.
User can determine lower/upper thresholds for gene and UMI counts to filter out.
}
