% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{mtRNA_plot_Preprocess}
\alias{mtRNA_plot_Preprocess}
\title{Calculate mitochondrial gene expression and plotting the distribution.}
\usage{
mtRNA_plot_Preprocess(Obj.flt, prefix, OutPath)
}
\arguments{
\item{Obj.flt}{A Seurat object where cells with too low gene/UMI counts or high gene counts are filtered out.}

\item{prefix}{Prefix of mitochondrial genes of the species of your data. (not required but recommended)}

\item{OutPath}{Path to the output directory, where its subdirectories contain log files generated in previous steps.}
}
\description{
Percentage of mitochondrial gene expression in each cell is calculated and plotted.
}
