% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{NormToPCA_Process}
\alias{NormToPCA_Process}
\title{Normalize the data, find highly variable genes, scale the data and perform PCA}
\usage{
NormToPCA_Process(Obj.flt, OutPath, num.dim = 50)
}
\arguments{
\item{Obj.flt}{A Seurat object where cells with too low gene/UMI counts or high gene counts are filtered out.}

\item{OutPath}{Path to the output directory, where its subdirectories contain log files generated in previous steps.}

\item{num.dim}{Number of principal components to calculate. Defaults to the largest available limit, 50.}
}
\description{
Normalization, Finding variable genes, scaling and PCA are performed.
}
