% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{Run_DoubletFinder}
\alias{Run_DoubletFinder}
\title{Running DoubletFinder (C. S. McGinnis et.al 2019)}
\usage{
Run_DoubletFinder(
  object,
  npc,
  nLib = 1,
  nRun = 1,
  rseed = 2022,
  est.rates,
  OutPath
)
}
\arguments{
\item{object}{A fully-processed (from QC to cell type annotation) Seurat object.}

\item{npc}{Number of principal components to use.}

\item{nLib}{The number of libraries (samples) of data to run. Default to 1}

\item{nRun}{The number of running iteration with different random seed. Default to 1}

\item{rseed}{Initial random seed to use for running DoubletFinder. Each of nRun iterations uses rseed multiplied by iteration number.}

\item{est.rates}{Vector containing estimated rate of doublets (multiplets) of the data, which is in range (0,1).}

\item{OutPath}{Path to the output directory, where its subdirectories contain log files generated in previous steps.}
}
\description{
Running DoubletFinder (C. S. McGinnis et.al 2019)
}
