% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Define_Doublets.R
\name{Scale_AvgScores}
\alias{Scale_AvgScores}
\title{Normalizing doublet scores between 0 and 1}
\usage{
Scale_AvgScores(type, dblobj, nRun = 10, nSet)
}
\arguments{
\item{type}{Type of the input dataset, either 'raw' or 'simul'.}

\item{dblobj}{A list of doublet scores, resulted from function Bind_Doublet_Scores().}

\item{nRun}{An integer indicating the number of running iteration with different random seeds for each dataset. Default to 10}

\item{nSet}{The number of datasets to run. Default to 1 for 'raw' and 100 for 'simul'.}
}
\description{
For each dataset and detection tool, doublet scores previously calculated in 'Bind_Doublet_Scores()'
are averaged and normalized to range (0,1) by min-max scaling.
}
