% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect_Doublets.R
\name{TimeCheck}
\alias{TimeCheck}
\title{Check Execution time}
\usage{
TimeCheck(start, end)
}
\arguments{
\item{start}{Execution start time}

\item{end}{Execution end time}
}
\description{
Check Execution time
}
