% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{UMAPtoDEG_Process}
\alias{UMAPtoDEG_Process}
\title{Running UMAP, Louvain algorithm for clustering and Finding differentially expressed genes (DEGs)}
\usage{
UMAPtoDEG_Process(Obj.norm, nPC = 20, genes, seed = 777778, res = 0.8, OutPath)
}
\arguments{
\item{Obj.norm}{A Seurat object where analysis processes are all performed in the previous step 'NormtoPCA_Process()'.}

\item{nPC}{Number of principal components to use. Default to 20.}

\item{genes}{A vector containing genes to use when plotting features.}

\item{seed}{Random seed number to use for generating UMAP plots. Default to '777778'}

\item{res}{Value of the resolution parameter which is proportional to the number of communities. Default to 0.8}

\item{OutPath}{Path to the output directory, where its subdirectories contain log files generated in previous steps.}
}
\description{
UMAP, clustering and DE analysis are done to the input data sequentially, a Seurat object on which PCA was previously performed .
}
