% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing.R
\name{Upper_Preprocess}
\alias{Upper_Preprocess}
\title{Filtering out cells based on predetermined lower/upper thresholds for gene/UMI counts.}
\usage{
Upper_Preprocess(
  rawObj,
  thres.gene,
  thres.umi,
  thres.UpGene,
  thres.prob = 0.9,
  OutPath
)
}
\arguments{
\item{rawObj}{A Seurat object created with input expression matrix.}

\item{thres.gene}{A numberic value indicating lower gene count threshold.}

\item{thres.umi}{A numberic value indicating lower UMI count threshold.}

\item{thres.UpGene}{A numberic value indicating upper gene count threshold. Either thres.UpGene or thres.prob should be supplied.}

\item{thres.prob}{Threshold for posterior probability of cells belonging to component-3 in the gene count mixture model.
Cells with their posterior probability higher than this value (Defaults to 0.7) are regarded as putative multiplets and filtered out.
Either thres.UpGene or thres.prob should be supplied.}

\item{OutPath}{Path to the output directory, where its subdirectories contain log files generated in previous steps.}
}
\description{
Cells with too low gene/UMI counts are filtered out based on user-defined thresholds (thres.gene, thres.umi) in previous step.
For upper gene count filtering, either thres.UpGene or thres.prob should be provided.
If user put in thres.prob (in range 0 to 1), parameters of Gaussian mixture model (number of components = 3)
fitted gene counts are estimated, and cells with higher posterior probability of belonging to component-3 than thres.prob (default=0.7)
are filtered out, being regarded as putative multiplets.
}
